Fix or silence (msvc) compiler warnings about unused formal parameters, unused local variables, narrowing conversions, constant conditional expressions, user-defined comma operators, name hiding, ambiguous overloads, and signed-unsigned-mismatches
(cherry picked from commit a5577b42342692c5ee7860f2d62c6ca49e476692)

+++ include/boost/graph/gursoy_atun_layout.hpp
@@ -318,7 +318,7 @@ void gursoy_atun_layout(const VertexListAndIncidenceGraph& graph,
                         const Topology& space,
                         PositionMap position)
 {
-  gursoy_atun_layout(graph, space, position, num_vertices(graph));
+  gursoy_atun_layout(graph, space, position, static_cast<int>(num_vertices(graph)));
 }
 
 template<typename VertexListAndIncidenceGraph, typename Topology,
@@ -337,7 +337,7 @@ gursoy_atun_layout(const VertexListAndIncidenceGraph& graph,
   std::pair<double, double> learn(0.8, 0.2);
   gursoy_atun_layout(graph, space, position,
                      choose_param(get_param(params, iterations_t()),
-                                  num_vertices(graph)),
+                                  static_cast<int>(num_vertices(graph))),
                      choose_param(get_param(params, diameter_range_t()), 
                                   diam).first,
                      choose_param(get_param(params, diameter_range_t()),
